requirements.txt
# Core ML
torch>=2.0.0
lightning>=2.0.0

# Data
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Alpaca API
alpaca-py>=0.21.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0

